Insert a load fence before s.cursor.get()
As per @dvyokov's comment here https://github.com/jakewins/4fq/issues/1

I'm unsure about the implications here - the algorithm is meant to
handle this read coming back as any random value. The problem is, as
outlined in that ticket, if you do things that moves you outside the
confines of what is defined, that makes the entire program effectively
undefined, which we don't want.

This does incur a major performance overhead, 5-8% in benchmarks. It may
be possible to review the Go compiler behavior and show that, while
theoretically possible, practically this is safe. However, that leaves
gccgo on the side.

Hence: Default to making this safe, and then come back and see if there
are reasonable ways to leave this fence out.