Make barrier an interface
The existing barrier impl is multi-writer, eg. it safely handles
multiple writers publishing on the barrier at the same time.

In order to build a multi-consumer queue, we want this same barrier
on the "other side", publishing which slots are free and ready for
new adventures.

I'm trying a path to that here where I want to first introduce
a single-writer barrier and slot that into the existing MPSC
queue. The theory then being that I could then simply swap out
the barrier impl to get a safe MPMC queue.