Replace simple consumer sequence with barrier
This makes no actual difference to the current
implementation - but it crucially sets the sequencer
up to wait on a barrier, rather than directly
on a sequence. This would then allow a multi-writer
barrier to build an multi-consumer queue.

There's something worthy of note here: This is
moving away a bit from the disruptor design. The barriers
in disruptor point to other data structures, they
don't themselves constitute places where you can
publish how far along a sequence you are.

I'm pretty sure the approach implemented here is
an affront to the intention of the name "Barrier" in
disruptor. I was thinking of it in terms of "a data structure
that forms a wall between threads working on different
sequence ranges.

In any case, this may all need refactoring again as I'm
learning the ins and outs here.