Add SPSC queue
This builds on the work in the prior commits today to replace the
multi-writer barrier in the mpsc queue with a single-writer barrier,
and slapping a nice "SPSC" sticker on it.

Benchmarks at ~70ns, which is likely mostly just because of the lowered
contention, but not having the overhead of the multi-writer publish
algo likely plays a big role as well.

Notably this still uses the CAS in sequencer. A simple step to improve
throughput further would be to add a specialized singleWriter sequencer
that does an ordered store instead.